Move TraceProtoUtils to otprotocol artifact.
