+++ exporters/otprotocol/README.md
@@ -1,0 +1,4 @@
+# OpenTelemetry Protocol Exporter - gRPC
+
+This is the OpenTelemetry exporter, sending span data to OpenTelemetry collector
+via gRPC.
