+++ exporters/otprotocol/build.gradle
@@ -1,0 +1,16 @@
+description = 'OpenTelemetry Protocol Exporter'
+
+dependencies {
+    api project(':opentelemetry-proto'),
+            project(':opentelemetry-sdk'),
+            libraries.grpc_protobuf,
+            libraries.grpc_stub,
+            libraries.protobuf,
+            libraries.protobuf_util
+
+    testImplementation "io.grpc:grpc-testing:${grpcVersion}"
+    testRuntime "io.grpc:grpc-netty-shaded:${grpcVersion}"
+
+    signature "org.codehaus.mojo.signature:java17:1.0@signature"
+    signature "net.sf.androidscents.signature:android-api-level-14:4.0_r4@signature"
+}
