+++ exporters/otprotocol/src/test/java/io/opentelemetry/exporters/otprotocol/TraceProtoUtilsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.exporters.otproto;
+package io.opentelemetry.exporters.otprotocol;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -25,7 +25,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link io.opentelemetry.exporters.otproto.TraceProtoUtils}. */
+/** Unit tests for {@link io.opentelemetry.exporters.otprotocol.TraceProtoUtils}. */
 @RunWith(JUnit4.class)
 public class TraceProtoUtilsTest {
